Support using existing Argo tunnels
Cloudflare recommended approach is to create the tunnels and then launch them in
Kubernetes by supplying the credentials file to the deployment. This way of
launching the tunnels is now supported.

The existing ability for the chart to create the tunnel remains as I believe it
is a useful way of testing tunnels quickly.